Get Correlation id from Operation parameters if available
