+++ common/src/main/java/com/microsoft/identity/common/internal/controllers/CommandDispatcher.java
@@ -25,13 +25,11 @@ package com.microsoft.identity.common.internal.controllers;
 import android.content.Intent;
 import android.os.Handler;
 import android.os.Looper;
+import android.text.TextUtils;
 import android.util.Pair;
-
-import androidx.annotation.NonNull;
 
 import com.microsoft.identity.common.exception.BaseException;
 import com.microsoft.identity.common.exception.UserCancelException;
-import com.microsoft.identity.common.internal.cache.ICacheRecord;
 import com.microsoft.identity.common.internal.logging.DiagnosticContext;
 import com.microsoft.identity.common.internal.logging.Logger;
 import com.microsoft.identity.common.internal.providers.oauth2.AuthorizationActivity;
@@ -41,10 +39,12 @@ import com.microsoft.identity.common.internal.result.AcquireTokenResult;
 import com.microsoft.identity.common.internal.result.ILocalAuthenticationResult;
 import com.microsoft.identity.common.internal.telemetry.Telemetry;
 
-import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 public class CommandDispatcher {
 
@@ -68,7 +68,7 @@ public class CommandDispatcher {
         sSilentExecutor.execute(new Runnable() {
             @Override
             public void run() {
-                final String correlationId = initializeDiagnosticContext();
+                final String correlationId = initializeDiagnosticContext(command.getParameters().getCorrelationId());
 
                 Object result = null;
                 BaseException baseException = null;
@@ -182,7 +182,9 @@ public class CommandDispatcher {
             sInteractiveExecutor.execute(new Runnable() {
                 @Override
                 public void run() {
-                    final String correlationId = initializeDiagnosticContext();
+                    final String correlationId = initializeDiagnosticContext(
+                            command.getParameters().getCorrelationId()
+                    );
 
                     if (command.getParameters() instanceof AcquireTokenOperationParameters) {
                         logInteractiveRequestParameters(methodName, (AcquireTokenOperationParameters) command.getParameters());
@@ -381,9 +383,13 @@ public class CommandDispatcher {
         }
     }
 
-    public static String initializeDiagnosticContext() {
+    public static String initializeDiagnosticContext(@Nullable final String requestCorrelationId) {
         final String methodName = ":initializeDiagnosticContext";
-        final String correlationId = UUID.randomUUID().toString();
+
+        final String correlationId = TextUtils.isEmpty(requestCorrelationId) ?
+                UUID.randomUUID().toString() :
+                requestCorrelationId;
+
         final com.microsoft.identity.common.internal.logging.RequestContext rc =
                 new com.microsoft.identity.common.internal.logging.RequestContext();
         rc.put(DiagnosticContext.CORRELATION_ID, correlationId);
